Make priority into SMALLINT

+++ db-scheduler/src/test/resources/mysql_tables.sql
@@ -10,7 +10,7 @@ create table test.scheduled_tasks (
   consecutive_failures INT,
   last_heartbeat timestamp(6) null,
   version BIGINT not null,
-  priority INT,
+  priority SMALLINT,
   PRIMARY KEY (task_name, task_instance),
   INDEX execution_time_idx (execution_time),
   INDEX last_heartbeat_idx (last_heartbeat),
