+++ db-scheduler/src/test/resources/postgresql_tables.sql
@@ -10,10 +10,14 @@ create table scheduled_tasks (
   consecutive_failures INT,
   last_heartbeat timestamp with time zone,
   version BIGINT not null,
-  priority INT,
+  priority SMALLINT,
   PRIMARY KEY (task_name, task_instance)
 );
 
 CREATE INDEX execution_time_idx ON scheduled_tasks (execution_time);
 CREATE INDEX last_heartbeat_idx ON scheduled_tasks (last_heartbeat);
 CREATE INDEX priority_execution_time_idx on scheduled_tasks (priority desc, execution_time asc);
+
+-- an optimization for users of priority might be to add priority to the execution_time_idx
+-- this _might_ save reads as the priority-value is already in the index
+-- CREATE INDEX execution_time_idx ON scheduled_tasks (execution_time asc, priority desc);
